\name{bootstrap}
\alias{bootstrap}
\title{bootstrap: a function to compute bootstrap replicates}
\usage{
  bootstrap(x, ...)
}
\arguments{
  \item{modelfit}{a multiOU object, (output of
  multiTypeOU() fn)}
}
\description{
  bootstrap: a function to compute bootstrap replicates
}
\details{
  define the generic S3 bootstrap method
}

