\name{convert}
\alias{convert}
\title{toggles between ouch and ape format trees}
\usage{
  convert(ot, regimes = NULL)
}
\arguments{
  \item{ot}{a phylogenetic tree of class "phylo" or
  "ouchtree"}

  \item{regimes}{optional list of regimes for an ouchtree}
}
\value{
  a phylogenetic tree in the alternative format
}
\description{
  toggles between ouch and ape format trees
}

