\name{get_clade}
\alias{get_clade}
\title{A function that returns the edge numbers of all branches
descended from the mrca of sp1 and sp2.}
\usage{
  get_clade(phy, sp1, sp2)
}
\arguments{
  \item{phy}{a phylogenetic tree of class phylo (ape)}

  \item{sp1}{the tip.label of the 1st species}

  \item{sp1}{the tip.label of the 2nd species the
  descendents of the MRCA of sp1 and sp2 define the regime}
}
\description{
  A function that returns the edge numbers of all branches
  descended from the mrca of sp1 and sp2.
}

