\name{getParameters.multiOU}
\alias{getParameters.multiOU}
\title{Internal, extends method to mulitOU class}
\usage{
  \method{getParameters}{multiOU} (ws)
}
\arguments{
  \item{a}{multiOU class object}
}
\value{
  a vector of all the model parameter estimates
}
\description{
  Internal, extends method to mulitOU class
}

