\name{llik.closure}
\alias{llik.closure}
\title{returns the likelihood function for the chosen submodel}
\usage{
  llik.closure(data, tree, regimes, model_spec,
    fixed = list(alpha = 1e-12), neg_llik = FALSE)
}
\description{
  returns the likelihood function for the chosen submodel
}
\keyword{internal}

