\name{loglik.multiOU}
\alias{loglik.multiOU}
\title{Internal use mostly extending the method to multiOU class
exported for use by PMC}
\usage{
  \method{loglik}{multiOU} (ws)
}
\arguments{
  \item{a}{multiOU class object (model fit)}
}
\value{
  extracts the log likelihood value
}
\description{
  Internal use mostly extending the method to multiOU class
  exported for use by PMC
}

