\name{mrcaOUCH}
\alias{mrcaOUCH}
\title{Finds the most recent common ancestor on an OUCH formatted tree}
\usage{
  mrcaOUCH(cladeVector, tree)
}
\arguments{
  \item{cladeVector}{vector of species for which you want
  to find the most recent common ancestor}

  \item{tree}{an ouchtree class phylogeny}
}
\value{
  the node number of the most recent common ancestor
}
\description{
  Finds the most recent common ancestor on an OUCH
  formatted tree
}

