\name{update.multiOU}
\alias{update.multiOU}
\title{updates a multiOU model parameter estimation on new data}
\usage{
  \method{update}{multiOU} (model, data)
}
\arguments{
  \item{model}{a multiOU model}

  \item{data}{trait data on which to base the update}
}
\description{
  updates a multiOU model parameter estimation on new data
}

