\name{wrightscape}
\alias{wrightscape}
\title{fit the full model entirely in C code, for speed.}
\usage{
  wrightscape(data, tree, regimes, alpha = 1, sigma = 1,
    theta = NULL, Xo = NULL, use_siman = 0)
}
\description{
  fit the full model entirely in C code, for speed.
}
\details{
  Does not allow fitting of the submodels.  Largely
  historical
}
\keyword{internal}

